commmti

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -16,8 +16,8 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 char *str1 = s1;
 char *str2 = s2;
 char *string;
-size_t cn = n;
-size_t len;
+unsigned int cn = n;
+unsigned int len;
 unsigned int k;
 
 if (str1 == NULL)
